preprocessing exploration of the data was  added
other files were adapted according to modifications for the data exploration using default settings.  Preprocessing exploration of the data was added to evaluate the effect of the cutoff thresholds proposed by the authors in our data. Improves in the filtering of the contigs equation were made. further results evaluation are seeing in the exploration file.